requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.7.2
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudscraper==1.2.71
colorama==0.4.6
coverage==7.5.0
distro==1.9.0
dj-database-url==2.3.0
django>=4.2.0
django-cors-headers>=4.3.1
django-redis>=5.3.0
django-ratelimit==4.1.0
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
djangorestframework-stubs==3.14.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
kombu==5.4.2
ndjson==0.3.1
openai==1.59.8
packaging==24.2
pillow==10.2.0
pluggy==1.5.0
prompt_toolkit==3.0.48
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.10.1
pyparsing==3.2.1
pytest==7.4.4
pytest-cov==6.0.0
pytest-django==4.9.0
pytest-mock==3.14.0
python-chess==1.999
python-dateutil==2.8.2
python-dotenv
redis>=4.5.4
pytest
pytest-django
pytest-cov
flake8
psycopg2-binary>=2.9.9
boto3>=1.34.7
django-storages>=1.14.2
gunicorn>=21.2.0
whitenoise>=6.6.0
sentry-sdk>=1.39.1
python-jose>=3.3.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
requests>=2.31.0
stockfish>=3.28.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==11.4.1
tqdm==4.67.1
types-PyYAML==6.0.12.20241230
types-requests==2.31.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
rq==1.15.1
tenacity>=8.2.3
# Telemetry dependencies
prometheus-client==0.19.0
psutil==5.9.8
django-prometheus==2.3.1
django-environ==0.12.0
